+++ src/components/AreaMap/style.less
@@ -1,3 +1,5 @@
+@import "../../containers/App/theme";
+
 .map-form{
   position: relative;
   z-index: 99;
@@ -23,7 +25,7 @@
     input{
       margin-top: 10px !important;
       font-size: 14px !important;
-      color: #5E6670 !important;
+      color: @dark-gray-color !important;
     }
     &.border-right{
       border-right: 1px solid #E4E4E4;
