+++ src/components/DataTable/style.less
@@ -50,6 +50,20 @@
   padding-left:10px;
   padding-right: 10px;
 }
+.table-total-number{
+  margin-top: 35px;
+  display: block;
+  text-align: left;
+  i{
+    position: relative;
+    top:3px;
+    font-size: 16px;
+    color:@normal-gray-color;
+    text-align: right;
+    margin-right: 5px;
+    margin-left: 5px;
+  }
+}
 .campaign-data-table{
   .ant-table-header{
     overflow: auto !important;
