+++ src/components/IranMap/style.less
@@ -13,7 +13,7 @@
   stroke-width: 0.5;
   &.selected {
     cursor: pointer;
-    fill: #5E6670;
+    fill: @dark-gray-color;
   }
 }
