+++ src/components/SelectList/style.less
@@ -158,7 +158,7 @@
           color: white;
         }
         &:hover {
-          background: #5E6670 !important;
+          background: @dark-gray-color !important;
         }
         background: #B6BBC2;
         height: @campaign-button-height !important;
