+++ src/components/SelectTag/style.less
@@ -12,7 +12,7 @@
   margin-top: -10px;
   font-size: 12px;
   font-weight: 500;
-  color: #5E6670 ;
+  color: @dark-gray-color ;
 }
 .show-tag {
   display: inline-block;
