+++ src/containers/App/color.less
@@ -8,6 +8,8 @@
 @highlight-color        : @red-6;
 @warning-color          : @yellow-6;
 @normal-color           : #d9d9d9;
+@normal-gray-color      : #88909B;
+@dark-gray-color        : #5E6670;
 
 // Color used by default to control hover and active backgrounds and for
 // alert info backgrounds.
@@ -153,8 +155,9 @@
 
 // Tag
 // --
-@tag-default-bg: #f3f3f3;
+@tag-default-bg   : #f3f3f3;
 @tag-default-color: @text-color;
+@tag-active-color : #66C395;
 
 // TimePicker color
 @time-picker-selected-bg: @background-color-base;
