+++ src/containers/Campaign/containers/CheckPublish/style.less
@@ -9,7 +9,7 @@
 .check-title{
   border-bottom: 1px solid #E4E6E9;
   width: 100%;
-  color: #5E6670;
+  color: @dark-gray-color;
   font-size: 14px;
   font-weight:500;
   margin-top: 35px;
