+++ src/containers/Campaign/containers/SelectPublisher/style.less
@@ -43,7 +43,7 @@
   width: 370px !important;
   .publisher-content {
     span {
-      color: #5E6670;
+      color: @dark-gray-color;
       font-size: 13px;
     }
   }
@@ -84,12 +84,17 @@
 }
 
 .pub-table-title {
-  display: flex;
+  display: inline-flex;
   align-items: center;
-  margin-top: 50px;
   height: 30px;
 }
 
+.data-table-wrapper {
+  margin-top: 50px;
+}
+.data-table-description{
+  display: flex;
+}
 .customize-table-modal {
   min-width: 370px;
   display: table;
@@ -111,13 +116,14 @@
     }
   }
 }
-.pub-switch-wrapper{
+
+.pub-switch-wrapper {
   border: 1px solid #E4E6E9;
   padding: 10px;
   margin-top: 25px;
-  span{
+  span {
     margin: 5px;
     font-size: 13px;
-    color:#5E6670;
+    color: @dark-gray-color;
   }
 }
