+++ src/containers/Campaign/containers/Type/style.less
@@ -77,7 +77,7 @@
         display: flex;
         align-items: center;
         span , i{
-          color:#88909B !important;
+          color:@normal-gray-color !important;
         }
         span{
           order:1;
