+++ src/containers/Campaign/containers/Upload/style.less
@@ -59,7 +59,7 @@
           color: white;
         }
         &:hover {
-          background: #5E6670 !important;
+          background: @dark-gray-color !important;
         }
         background: #B6BBC2;
         height: @campaign-button-height !important;
@@ -76,7 +76,7 @@
   padding: 2px;
   margin: 2px 3px;
   background-color: #fff;
-  color: #88909B;
+  color: @normal-gray-color;
   box-shadow: 0 0 0 1px #d9d9d9 inset;
 }
 
@@ -238,7 +238,8 @@
   }
 }
 .modal-btn-row{
-  float: left;
+  display: block;
+  text-align: left;
 }
 .modal-switch-row {
   margin-bottom: 15px;
