+++ src/containers/Campaign/style.less
@@ -179,7 +179,7 @@ span[role="menuitem"] {
 
 .currency {
   font-size: 14px;
-  color: #5E6670;
+  color: @dark-gray-color;
 }
 
 // Datepicker 2 style
