+++ src/containers/User/containers/Login/style.less
@@ -29,7 +29,7 @@
 //Back button on login class Can be Removed
 .back-button{
   display: block;
-  color: #88909B;
+  color: @normal-gray-color;
   margin: 0;
 }
 .button-login-wrapper{
@@ -60,7 +60,7 @@
   top: 2px;
 }
 .back-arrow{
-  color: #88909B;
+  color: @normal-gray-color;
   margin-left: 3px;
   font-size: 10px;
   display: block;
